refactor(admin): decompose locations catalog page (REF-303)
- Extract LocationFormFields component
- Extract LocationDeleteModal component
- Reduce locations/page.tsx from 544 to 337 LOC
- Preserve page routing, search params, and catalog template orchestration
